+++ src/backend/langflow/graph/__init__.py
@@ -1,4 +1,35 @@
-from langflow.graph.base import Edge, Node
-from langflow.graph.graph import Graph
+from langflow.graph.edge.base import Edge
+from langflow.graph.graph.base import Graph
+from langflow.graph.node.base import Node
+from langflow.graph.node.types import (
+    AgentNode,
+    ChainNode,
+    DocumentLoaderNode,
+    EmbeddingNode,
+    LLMNode,
+    MemoryNode,
+    PromptNode,
+    TextSplitterNode,
+    ToolNode,
+    ToolkitNode,
+    VectorStoreNode,
+    WrapperNode,
+)
 
-__all__ = ["Graph", "Node", "Edge"]
+__all__ = [
+    "Graph",
+    "Node",
+    "Edge",
+    "AgentNode",
+    "ChainNode",
+    "DocumentLoaderNode",
+    "EmbeddingNode",
+    "LLMNode",
+    "MemoryNode",
+    "PromptNode",
+    "TextSplitterNode",
+    "ToolNode",
+    "ToolkitNode",
+    "VectorStoreNode",
+    "WrapperNode",
+]
