+++ src/backend/langflow/graph/graph/base.py
@@ -1,38 +1,20 @@
 from typing import Dict, List, Type, Union
 
-from langflow.graph.base import Edge, Node
-from langflow.graph.nodes import (
-    AgentNode,
-    ChainNode,
-    DocumentLoaderNode,
-    EmbeddingNode,
+from langflow.graph.edge.base import Edge
+from langflow.graph.graph.constants import NODE_TYPE_MAP
+from langflow.graph.node.base import Node
+from langflow.graph.node.types import (
     FileToolNode,
     LLMNode,
-    MemoryNode,
-    PromptNode,
-    TextSplitterNode,
     ToolkitNode,
-    ToolNode,
-    VectorStoreNode,
-    WrapperNode,
 )
-from langflow.interface.agents.base import agent_creator
-from langflow.interface.chains.base import chain_creator
-from langflow.interface.document_loaders.base import documentloader_creator
-from langflow.interface.embeddings.base import embedding_creator
-from langflow.interface.llms.base import llm_creator
-from langflow.interface.memories.base import memory_creator
-from langflow.interface.prompts.base import prompt_creator
-from langflow.interface.text_splitters.base import textsplitter_creator
-from langflow.interface.toolkits.base import toolkits_creator
-from langflow.interface.tools.base import tool_creator
 from langflow.interface.tools.constants import FILE_TOOLS
-from langflow.interface.vector_store.base import vectorstore_creator
-from langflow.interface.wrappers.base import wrapper_creator
 from langflow.utils import payload
 
 
 class Graph:
+    """A class representing a graph of nodes and edges."""
+
     def __init__(
         self,
         nodes: List[Dict[str, Union[str, Dict[str, Union[str, List[str]]]]]],
@@ -43,6 +25,7 @@ class Graph:
         self._build_graph()
 
     def _build_graph(self) -> None:
+        """Builds the graph from the nodes and edges."""
         self.nodes = self._build_nodes()
         self.edges = self._build_edges()
         for edge in self.edges:
@@ -51,17 +34,25 @@ class Graph:
 
         # This is a hack to make sure that the LLM node is sent to
         # the toolkit node
+        self._build_node_params()
+        # remove invalid nodes
+        self._remove_invalid_nodes()
+
+    def _build_node_params(self) -> None:
+        """Identifies and handles the LLM node within the graph."""
         llm_node = None
         for node in self.nodes:
             node._build_params()
-
             if isinstance(node, LLMNode):
                 llm_node = node
 
-        for node in self.nodes:
-            if isinstance(node, ToolkitNode):
-                node.params["llm"] = llm_node
-        # remove invalid nodes
+        if llm_node:
+            for node in self.nodes:
+                if isinstance(node, ToolkitNode):
+                    node.params["llm"] = llm_node
+
+    def _remove_invalid_nodes(self) -> None:
+        """Removes invalid nodes from the graph."""
         self.nodes = [
             node
             for node in self.nodes
@@ -70,19 +61,23 @@ class Graph:
         ]
 
     def _validate_node(self, node: Node) -> bool:
+        """Validates a node."""
         # All nodes that do not have edges are invalid
         return len(node.edges) > 0
 
     def get_node(self, node_id: str) -> Union[None, Node]:
+        """Returns a node by id."""
         return next((node for node in self.nodes if node.id == node_id), None)
 
     def get_nodes_with_target(self, node: Node) -> List[Node]:
+        """Returns the nodes connected to a node."""
         connected_nodes: List[Node] = [
             edge.source for edge in self.edges if edge.target == node
         ]
         return connected_nodes
 
     def build(self) -> List[Node]:
+        """Builds the graph."""
         # Get root node
         root_node = payload.get_root_node(self)
         if root_node is None:
@@ -90,6 +85,7 @@ class Graph:
         return root_node.build()
 
     def get_node_neighbors(self, node: Node) -> Dict[Node, int]:
+        """Returns the neighbors of a node."""
         neighbors: Dict[Node, int] = {}
         for edge in self.edges:
             if edge.source == node:
@@ -105,6 +101,7 @@ class Graph:
         return neighbors
 
     def _build_edges(self) -> List[Edge]:
+        """Builds the edges of the graph."""
         # Edge takes two nodes as arguments, so we need to build the nodes first
         # and then build the edges
         # if we can't find a node, we raise an error
@@ -121,30 +118,15 @@ class Graph:
         return edges
 
     def _get_node_class(self, node_type: str, node_lc_type: str) -> Type[Node]:
-        node_type_map: Dict[str, Type[Node]] = {
-            **{t: PromptNode for t in prompt_creator.to_list()},
-            **{t: AgentNode for t in agent_creator.to_list()},
-            **{t: ChainNode for t in chain_creator.to_list()},
-            **{t: ToolNode for t in tool_creator.to_list()},
-            **{t: ToolkitNode for t in toolkits_creator.to_list()},
-            **{t: WrapperNode for t in wrapper_creator.to_list()},
-            **{t: LLMNode for t in llm_creator.to_list()},
-            **{t: MemoryNode for t in memory_creator.to_list()},
-            **{t: EmbeddingNode for t in embedding_creator.to_list()},
-            **{t: VectorStoreNode for t in vectorstore_creator.to_list()},
-            **{t: DocumentLoaderNode for t in documentloader_creator.to_list()},
-            **{t: TextSplitterNode for t in textsplitter_creator.to_list()},
-        }
-
+        """Returns the node class based on the node type."""
         if node_type in FILE_TOOLS:
             return FileToolNode
-        if node_type in node_type_map:
-            return node_type_map[node_type]
-        if node_lc_type in node_type_map:
-            return node_type_map[node_lc_type]
-        return Node
+        if node_type in NODE_TYPE_MAP:
+            return NODE_TYPE_MAP[node_type]
+        return NODE_TYPE_MAP[node_lc_type] if node_lc_type in NODE_TYPE_MAP else Node
 
     def _build_nodes(self) -> List[Node]:
+        """Builds the nodes of the graph."""
         nodes: List[Node] = []
         for node in self._nodes:
             node_data = node["data"]
@@ -157,6 +139,7 @@ class Graph:
         return nodes
 
     def get_children_by_node_type(self, node: Node, node_type: str) -> List[Node]:
+        """Returns the children of a node based on the node type."""
         children = []
         node_types = [node.data["type"]]
         if "node" in node.data:
